Refactor with interface and descriptions
