+++ src/components/ui/radioGroup/RadioGroupSlotRecipe.tsx
@@ -1,6 +1,9 @@
 import { radioGroupAnatomy } from '@ark-ui/anatomy';
 import { defineSlotRecipe } from '@pandacss/dev';
 
+/**
+ * Update styles for Avatar component.
+ */
 export const radioGroupSlotRecipe = defineSlotRecipe({
   className: 'radioGroup',
   slots: radioGroupAnatomy.keys(),
